backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.2.0
redis==5.0.1
httpx==0.26.0
PyYAML==6.0.1
pdfplumber==0.10.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
email-validator==2.1.0.post1
pytz==2023.3.post1
openpyxl==3.1.5
et_xmlfile==2.0.0
pytest==8.0.0
pytest-asyncio==0.23.5
